+++ src/EngineAll.h
@@ -20,6 +20,8 @@ EngineBase* CreateEngineFb2FromFile(Str fileName);
 EngineBase* CreateEngineFb2FromData(Str data);
 EngineBase* CreateEngineMobiFromFile(Str fileName);
 EngineBase* CreateEngineMobiFromData(Str data);
+Str ExtractPdfFromPrintReplicaFile(Str path);
+Str ExtractPdfFromPrintReplicaData(Str data);
 EngineBase* CreateEnginePdbFromFile(Str fileName);
 EngineBase* CreateEngineChmFromFile(Str fileName);
 EngineBase* CreateEngineHtmlFromFile(Str fileName);